+++ NEWS.md
@@ -1,3 +1,7 @@
+# labelled dev
+
+* `to_factor()`, `to_character()` and `to_labelled.factor()` now preserves variable label.
+
 # labelled 0.2.0
 
 * Following evolution of `haven`, `labelled` deosn't support missing values anymore (cf. https://github.com/hadley/haven/commit/4b12ff9d51ddb9e7486966b85e0bcff44992904d)